# specs/007-client-management-heartbeat/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: MAGI Device Client Management Contract
  version: 0.1.0
  description: >
    Contract for account-owned device clients, automatic TV registration,
    rotating refresh tokens, and foreground heartbeat. RFC 8628 authorization
    remains available for legacy clients; existing open read operations and
    confidential Client Credentials remain backward compatible.
servers:
  - url: /
...
paths:
...
  /api/account/device-authorizations/{userCode}:
    get:
      tags: [Account Clients]
      operationId: inspectDeviceAuthorization
      summary: Preview a pending device before approval
      security:
        - cookieAuth: []
      parameters:
        - $ref: "#/components/parameters/UserCode"
      responses:
        "200":
          description: Pending device summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeviceAuthorizationPreviewEnvelope"
        "401":
          $ref: "#/components/responses/SessionRequired"
        "404":
          $ref: "#/components/responses/AuthorizationCodeUnavailable"
        "429":
          $ref: "#/components/responses/RateLimited"
  /api/account/device-authorizations/{userCode}/approve:
    post:
      tags: [Account Clients]
      operationId: approveDeviceAuthorization
      summary: Bind the pending TV authorization to the logged-in account
      security:
        - cookieAuth: []
      parameters:
        - $ref: "#/components/parameters/UserCode"
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ApproveDeviceAuthorizationRequest"
      responses:
        "200":
          description: Approval recorded; TV may exchange its device code
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeviceAuthorizationDecisionEnvelope"
        "400":
          $ref: "#/components/responses/ValidationProblem"
        "401":
          $ref: "#/components/responses/SessionRequired"
        "404":
          $ref: "#/components/responses/AuthorizationCodeUnavailable"
        "409":
          $ref: "#/components/responses/AuthorizationAlreadyDecided"
        "429":
          $ref: "#/components/responses/RateLimited"
  /api/account/device-authorizations/{userCode}/deny:
    post:
      tags: [Account Clients]
      operationId: denyDeviceAuthorization
      summary: Deny a pending TV authorization
      security:
        - cookieAuth: []
      parameters:
        - $ref: "#/components/parameters/UserCode"
        - $ref: "#/components/parameters/IdempotencyKey"
      responses:
        "200":
          description: Denial recorded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeviceAuthorizationDecisionEnvelope"
        "401":
          $ref: "#/components/responses/SessionRequired"
        "404":
          $ref: "#/components/responses/AuthorizationCodeUnavailable"
        "409":
          $ref: "#/components/responses/AuthorizationAlreadyDecided"
        "429":
          $ref: "#/components/responses/RateLimited"
  /api/account/clients:
    get:
      tags: [Account Clients]
      operationId: listAccountDeviceClients
      summary: List only device clients owned by the logged-in account
      security:
        - cookieAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: pageSize
          in: query
          schema:
            type: integer
            enum: [20, 50, 100]
            default: 20
      responses:
        "200":
          description: Stable presence-sorted page
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeviceClientPageEnvelope"
        "400":
          $ref: "#/components/responses/ValidationProblem"
        "401":
          $ref: "#/components/responses/SessionRequired"
  /api/account/clients/{clientId}:
    patch:
      tags: [Account Clients]
      operationId: renameAccountDeviceClient
      summary: Rename an owned active device client
      security:
        - cookieAuth: []
      parameters:
        - $ref: "#/components/parameters/ClientId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RenameDeviceClientRequest"
      responses:
        "200":
          description: Device renamed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeviceClientEnvelope"
        "400":
          $ref: "#/components/responses/ValidationProblem"
        "401":
          $ref: "#/components/responses/SessionRequired"
        "404":
          $ref: "#/components/responses/ClientNotFound"
        "409":
          description: Revoked clients cannot be renamed
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
  /api/account/clients/{clientId}/revoke:
    post:
      tags: [Account Clients]
      operationId: revokeAccountDeviceClient
      summary: Irreversibly revoke an owned device and every associated token
      security:
        - cookieAuth: []
      parameters:
        - $ref: "#/components/parameters/ClientId"
        - $ref: "#/components/parameters/IdempotencyKey"
      responses:
        "200":
          description: Revoked, or already revoked by an identical retry
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RevokeDeviceClientEnvelope"
        "401":
          $ref: "#/components/responses/SessionRequired"
        "404":
          $ref: "#/components/responses/ClientNotFound"
        "409":
          description: The same idempotency key was reused with a different request
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: opaque
    cookieAuth:
      type: apiKey
      in: cookie
      name: better-auth.session_token
  parameters:
    UserCode:
      name: userCode
      in: path
      required: true
      schema:
        type: string
        pattern: "^[A-HJ-NP-Z2-9]{4}-?[A-HJ-NP-Z2-9]{4}$"
      description: Short-lived code shown on the TV; normalized case-insensitively
    ClientId:
      name: clientId
      in: path
      required: true
      schema:
        type: string
        format: uuid
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      schema:
        type: string
        minLength: 1
        maxLength: 120
  schemas:
...
    DeviceAuthorizationPreview:
      type: object
      additionalProperties: false
      required:
        - userCode
        - deviceType
        - platform
        - platformVersion
        - appVersion
        - identitySummary
        - expiresAt
      properties:
        userCode:
          type: string
        deviceType:
          $ref: "#/components/schemas/DeviceType"
        platform:
          type: string
        platformVersion:
          type: string
        appVersion:
          type: string
        identitySummary:
          type: string
        suggestedName:
          type: [string, "null"]
        expiresAt:
          type: string
          format: date-time
    ApproveDeviceAuthorizationRequest:
      type: object
      additionalProperties: false
      required: [displayName]
      properties:
        displayName:
          $ref: "#/components/schemas/DisplayName"
    DeviceAuthorizationDecision:
      type: object
      additionalProperties: false
      required: [userCode, status, expiresAt]
      properties:
        userCode:
          type: string
        status:
          type: string
          enum: [approved, denied]
        expiresAt:
          type: string
          format: date-time
    DeviceType:
      type: string
      enum: [android_tv]
    DisplayName:
      type: string
      minLength: 1
      maxLength: 64
      description: Trimmed printable Unicode; duplicate names are allowed
    DeviceClientStatus:
      type: string
      enum: [online, offline, revoked]
    DeviceClient:
      type: object
      additionalProperties: false
      required:
        - id
        - displayName
        - deviceType
        - platform
        - platformVersion
        - appVersion
        - identitySummary
        - status
        - registeredAt
      properties:
        id:
          type: string
          format: uuid
        displayName:
          $ref: "#/components/schemas/DisplayName"
        deviceType:
          $ref: "#/components/schemas/DeviceType"
        platform:
          type: string
        platformVersion:
          type: string
        appVersion:
          type: string
        identitySummary:
          type: string
        status:
          $ref: "#/components/schemas/DeviceClientStatus"
        registeredAt:
          type: string
          format: date-time
        lastActiveAt:
          type: [string, "null"]
          format: date-time
        revokedAt:
          type: [string, "null"]
          format: date-time
    DeviceClientPage:
      type: object
      additionalProperties: false
      required: [items, total, page, pageSize, totalPages, asOf]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/DeviceClient"
        total:
          type: integer
          minimum: 0
        page:
          type: integer
          minimum: 1
        pageSize:
          type: integer
        totalPages:
          type: integer
          minimum: 1
        asOf:
          type: string
          format: date-time
          description: Single database time used for presence derivation
    RenameDeviceClientRequest:
      type: object
      additionalProperties: false
      required: [displayName]
      properties:
        displayName:
          $ref: "#/components/schemas/DisplayName"
    RevokeDeviceClientResult:
      type: object
      additionalProperties: false
      required: [client, accessTokensRevoked, refreshTokensRevoked]
      properties:
        client:
          $ref: "#/components/schemas/DeviceClient"
        accessTokensRevoked:
          type: integer
          minimum: 0
        refreshTokensRevoked:
          type: integer
          minimum: 0
    Problem:
      type: object
      additionalProperties: true
      required: [code, title, status]
      properties:
        code:
          type: string
        title:
          type: string
        status:
          type: integer
        detail:
          type: [string, object, "null"]
        requestId:
          type: [string, "null"]
          maxLength: 120
...
    DeviceAuthorizationPreviewEnvelope:
      $ref: "#/components/schemas/SuccessEnvelopeAuthorizationPreview"
    DeviceAuthorizationDecisionEnvelope:
      $ref: "#/components/schemas/SuccessEnvelopeAuthorizationDecision"
    DeviceClientEnvelope:
      $ref: "#/components/schemas/SuccessEnvelopeDeviceClient"
    DeviceClientPageEnvelope:
      $ref: "#/components/schemas/SuccessEnvelopeDeviceClientPage"
    RevokeDeviceClientEnvelope:
      $ref: "#/components/schemas/SuccessEnvelopeRevokeDeviceClient"
...
    SuccessEnvelopeAuthorizationPreview:
      type: object
      additionalProperties: false
      required: [success, data]
      properties:
        success:
          type: boolean
          const: true
        data:
          $ref: "#/components/schemas/DeviceAuthorizationPreview"
    SuccessEnvelopeAuthorizationDecision:
      type: object
      additionalProperties: false
      required: [success, data]
      properties:
        success:
          type: boolean
          const: true
        data:
          $ref: "#/components/schemas/DeviceAuthorizationDecision"
    SuccessEnvelopeDeviceClient:
      type: object
      additionalProperties: false
      required: [success, data]
      properties:
        success:
          type: boolean
          const: true
        data:
          $ref: "#/components/schemas/DeviceClient"
    SuccessEnvelopeDeviceClientPage:
      type: object
      additionalProperties: false
      required: [success, data]
      properties:
        success:
          type: boolean
          const: true
        data:
          $ref: "#/components/schemas/DeviceClientPage"
    SuccessEnvelopeRevokeDeviceClient:
      type: object
      additionalProperties: false
      required: [success, data]
      properties:
        success:
          type: boolean
          const: true
        data:
          $ref: "#/components/schemas/RevokeDeviceClientResult"
  responses:
    ValidationProblem:
      description: Request validation failed
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
...
    SessionRequired:
      description: Valid better-auth session required
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
    AuthorizationCodeUnavailable:
      description: Code is unknown, expired, consumed, or unavailable
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
    AuthorizationAlreadyDecided:
      description: A conflicting approval or denial already won
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
    ClientNotFound:
      description: Client does not exist or is owned by another account
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
    RateLimited:
      description: Request rate exceeded
      headers:
        Retry-After:
          required: true
          schema:
            type: integer
            minimum: 1
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"